+++ College04/sem1/Programming/25029344 Arjan Upadhayaya Timalsina/PersonalPlan.java
@@ -6,7 +6,7 @@ public class PersonalPlan extends AIModel
   /**
    * Constructor for objects of class PersonalPlan
    */
-  public PersonalPlan(String modelName, double price, long parameterCount, String contextWindowSize , int remainingPromptsinMonth)
+  public PersonalPlan(String modelName, double price, int parameterCount, String contextWindowSize , int remainingPromptsinMonth)
   {
     // initialise instance variables
     super(modelName, price, parameterCount, contextWindowSize);
